requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.38.46
botocore==1.38.46
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
decorator==4.4.2
distro==1.9.0
groq==0.29.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jmespath==1.0.1
moviepy==1.0.3
mysql-connector-python==9.3.0
numpy==2.3.1
opencv-python==4.11.0.86
pandas==2.3.0
pillow==11.2.1
proglog==0.1.12
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
